code cleanup, inconsistent line endings, performance improvements, migrations JDK5,7,8 to JDK11

+++ src/test/java/com/rapiddweller/benerator/test/JSR303ConstraintValidatorMock.java
@@ -62,7 +62,7 @@ public class JSR303ConstraintValidatorMock extends AbstractConstraintValidator<A
 
     @Override
 	public boolean isValid(Integer candidate, ConstraintValidatorContext context) {
-    	return (candidate != null && value == candidate.intValue());
+    	return (candidate != null && value == candidate);
     }
 
 }
